% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_latlon_ports.R
\name{assign_latlon_ports}
\alias{assign_latlon_ports}
\title{Assign lat and long to ports}
\usage{
assign_latlon_ports(dat, saveToFile = F)
}
\arguments{
\item{dat}{data frame. This is output from process_data}

\item{saveToFile}{boolean. Save data set to rds file}
}
\value{
NULL
files are written to data-raw
\enumerate{
\item clean PORT Table with lat an lons
\item REVENUE file for ports within NEUS footprint
\item REVENUE flle for ports Outside NEUS footprint
}
}
\description{
We use PORTID, PORTNM, STATE in the data and cdfbs.port table
}
\details{
Need to deal with Recreational data
}
