% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_species_codes.R
\name{assign_species_codes}
\alias{assign_species_codes}
\title{assign species codes to atlantis groups}
\usage{
assign_species_codes(commercialData)
}
\description{
Reads species from dev branch of neus-atlantis
}
\seealso{
Other processData: 
\code{\link{assign_latlon_ports}()},
\code{\link{assign_squid_trips}()},
\code{\link{assign_unclassified_groups}()},
\code{\link{process_data}()}
}
\concept{processData}
