% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_squid_trips.R
\name{assign_squid_trips}
\alias{assign_squid_trips}
\title{Define trips targeting squid}
\usage{
assign_squid_trips(allData, p = 0.3)
}
\arguments{
\item{allData}{Data frame}

\item{p}{Numeric. Value used to define the proportion of squid (by landings) in a trip to identify as a squid trip}
}
\description{
Define trips as squid trips based on the percentage of squid caught in each trip
A new column is created containing a boolean
}
\seealso{
Other processData: 
\code{\link{assign_latlon_ports}()},
\code{\link{assign_species_codes}()},
\code{\link{assign_unclassified_groups}()},
\code{\link{process_data}()}
}
\concept{processData}
