% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_unclassified_groups.R
\name{assign_unclassified_groups}
\alias{assign_unclassified_groups}
\title{assign unclassified groups to atlantis groups}
\usage{
assign_unclassified_groups(allData, lbstotons = 2204.62)
}
\arguments{
\item{allData}{Data frame}
}
\description{
Reads species from dev branch of neus-atlantis. There are several codes in the landings data that
refer to unclassified species. These need to be assigned to functional groups
Only species remaining unclassified are unknown sharks
}
