% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_data_with_forcing.R
\name{compare_data}
\alias{compare_data}
\title{Compare landing from Geret to data used to calibrate atlantis}
\usage{
compare_data()
}
\description{
Compare landing from Geret to data used to calibrate atlantis
}
