% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_atlantis_groundfish_data.R
\name{create_atlantis_groundfish_data}
\alias{create_atlantis_groundfish_data}
\title{get processed groundfish data for atlantis use}
\usage{
create_atlantis_groundfish_data()
}
\value{
data frames
}
\description{
Reads in groundfish data and returns effort and landings by box for main ports
}
