% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_atlantis_scallop_data.R
\name{create_atlantis_scallop_data}
\alias{create_atlantis_scallop_data}
\title{get processed scallop data for atlantis use}
\usage{
create_atlantis_scallop_data()
}
\value{
data frames
}
\description{
Reads in scallop data and returns effort and landings by box for main ports
}
