% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataset.R
\name{create_dataset}
\alias{create_dataset}
\title{Filter/create data sets for use in vignettes}
\usage{
create_dataset(fleetType)
}
\arguments{
\item{fleetType}{character (scallop, groundfish)}
}
\description{
Read in the fully processed data and filters by gear type
}
