% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_vignette_dataset.R
\name{create_vignette_dataset}
\alias{create_vignette_dataset}
\title{Filter/create data sets for use in vignettes}
\usage{
create_vignette_dataset(fleetType)
}
\arguments{
\item{fleetType}{character (scallop, groundfish)}
}
\value{
list of 3 data frames
\item{neus}{commercial data from states within NEUS footprint}
\item{outside}{commercial data from states outside NEUS footprint}
\item{rec}{recreational data from all states}
}
\description{
Read in the fully processed data and filter by gear type
}
