% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_missing_ports.R
\name{fix_missing_ports}
\alias{fix_missing_ports}
\title{Fix missing ports or incomplete ports}
\usage{
fix_missing_ports(channel)
}
\description{
Some ports are not defined or have missing state or county.
However by looking at the permit holder and where they land you can identify some ports
}
