% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_effort.R
\name{get_effort}
\alias{get_effort}
\title{get effort by port, box, fleet for data sets}
\usage{
get_effort(fleetData, ports, combine)
}
\arguments{
\item{fleetData}{data frame. Data which contains the effort by box, and fleet data}

\item{ports}{vector, list of ports, by PORTID for which effort is requred}
}
\value{
data frame of port effort aggregated by year, box
}
\description{
Effort is measured as Days at Sea at the trip level
}
\examples{
\dontrun{
fleetData <- readRDS(here::here("data/NEGroundfishDataCAMS.rds"))
get_effort(fleetData,
           ports=c(220101, 320201, 240207, 240115, 240301, 240403, 420209, 350635, 320901, 240813, 240601),
           combine = data.frame(main = c(320201, 240115, 240301), associated = c(320901,240813, 240601) ))
}
}
