% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_effort_landings.R
\name{get_effort_landings}
\alias{get_effort_landings}
\title{get effort by port, box, fleet for data sets}
\usage{
get_effort_landings(fleetData, ports, combine, speciesCodes, lbs = T)
}
\arguments{
\item{fleetData}{data frame. Data which contains the effort by box, and fleet data. Make sure you understand the units of the landings}

\item{ports}{vector, list of ports, by PORTID for which effort is requred}

\item{combine}{data frame. main ports and associated ports (one to one)}

\item{speciesCodes}{character vector. Atlantis species codes}

\item{lbs}{Boolean. Are landings in lbs? (Default = T)}
}
\value{
data frame of port effort and landings aggregated by year, box, Gear category, species.
if in lbs, landings are converted to metric tons
}
\description{
Effort is measured as Days at Sea at the trip level
}
\examples{
\dontrun{
fleetData <- readRDS(here::here("data/NEGroundfishDataCAMS.rds"))
get_effort_landings2(fleetData$neus,
           ports=c(240403,240207,220101,240115,420209,240301,320201,350635),
           combine = data.frame(main = c(320201, 240115, 240301), associated = c(320901,240813, 240601) ),
           speciesCodes = c("COD","HAD","YTF","POL","PLA","WTF","WHK","WIF","RED","HAL","WPF","OPT","WOL"),
           lbs = T)
}
}
\seealso{
Other modelData: 
\code{\link{create_atlantis_groundfish_data}()},
\code{\link{create_atlantis_scallop_data}()}
}
\concept{modelData}
