% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_port_lat_lon.R
\name{get_port_lat_lon}
\alias{get_port_lat_lon}
\title{Pull port lat and long address}
\usage{
get_port_lat_lon(portData)
}
\arguments{
\item{portData}{Data frame. Each row is Portname, County, State}
}
\value{
Data frame with lat and lon appended
}
\description{
Uses tidygeocoder
}
