% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nafo.R
\name{plot_nafo}
\alias{plot_nafo}
\title{Plot NEUS footprint along with NAFO areas}
\usage{
plot_nafo()
}
\description{
Plot NEUS footprint along with NAFO areas
}
\seealso{
Other supporting: 
\code{\link{pull_port_tables}()},
\code{\link{read_data}()}
}
\concept{supporting}
