% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_port_tables.r
\name{pull_port_tables}
\alias{pull_port_tables}
\title{Create PortTables}
\usage{
pull_port_tables(channel, saveToFile = F)
}
\arguments{
\item{channel}{A connection object from dbutils::connect_to_database`}

\item{write}{Boolean. Write output to file}
}
\description{
From VTR and CFDBS
}
\seealso{
Other supporting: 
\code{\link{plot_nafo}()},
\code{\link{read_data}()}
}
\concept{supporting}
