% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read in processed fleet data}
\usage{
read_data()
}
\value{
list of commercial and rec data
}
\description{
Processing in process_data function
}
\seealso{
Other supporting: 
\code{\link{plot_nafo}()},
\code{\link{pull_port_tables}()}
}
\concept{supporting}
