% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_atlantis_groups.R
\name{update_atlantis_groups}
\alias{update_atlantis_groups}
\title{Update Atlantis Species groups using DMIS data}
\usage{
update_atlantis_groups()
}
\description{
Update Atlantis Species groups using DMIS data
}
