% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_effort_landings.R
\name{get_effort_landings}
\alias{get_effort_landings}
\title{get effort by port, box, fleet for data sets}
\usage{
get_effort_landings(fleetData, ports, combine, speciesCodes, lbs = T)
}
\arguments{
\item{fleetData}{data frame. Data which contains the effort by box, and fleet data. Make sure you understand the units of the landings}

\item{ports}{vector, list of ports, by PORTID for which effort is requred}

\item{combine}{data frame. main ports and associated ports (one to one)}

\item{speciesCodes}{character vector. Atlantis species codes}

\item{lbs}{Boolean. Are landings in lbs? (Default = T)}
}
\value{
data frame of port effort aggregated by year, box
}
\description{
Effort is measured as Days at Sea at the trip level
}
\examples{
\dontrun{
fleetData <- readRDS(here::here("data/NEGroundfishDataCAMS.rds"))
get_effort(fleetData,
           ports=c(220101, 320201, 240207, 240115, 240301, 240403, 420209, 350635, 320901, 240813, 240601),
           combine = data.frame(main = c(320201, 240115, 240301), associated = c(320901,240813, 240601) ))
}
}
\seealso{
Other modelData: 
\code{\link{create_atlantis_groundfish_data}()},
\code{\link{create_atlantis_scallop_data}()},
\code{\link{get_effort_landings2}()}
}
\concept{modelData}
